music-faker/requirements.txt
kafka-python>=2.0.2
faker>=19.0.0
pandas>=2.0.0
numpy>=1.24.0
flask>=2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
hdfs3>=0.3.1
